fix: cursor gets out of sync occasionally at image previewing through IIP under tmux
